// src/client/types.gen.ts
export const AppSchemaEmpresasSchemaSortOp = {
  ASC: 'ASC',
  DESC: 'DESC',
} as const

/**
 * SortOp
 */
export type AppSchemaEmpresasSchemaSortOp =
  (typeof AppSchemaEmpresasSchemaSortOp)[keyof typeof AppSchemaEmpresasSchemaSortOp]

/**
 * CursorDirection
 */
export const AppSchemaEmpresasSchemaCursorDirection = {
  NEXT: 'next',
  PREV: 'prev',
} as const

/**
 * CursorDirection
 */
export type AppSchemaEmpresasSchemaCursorDirection =
  (typeof AppSchemaEmpresasSchemaCursorDirection)[keyof typeof AppSchemaEmpresasSchemaCursorDirection]

/**
 * UsuariosInclude
 */
export const UsuariosInclude = {
  NOME_USUARIO: 'nome_usuario',
  NOME_COMPLETO: 'nome_completo',
  NIVEL: 'nivel',
  TELEFONE: 'telefone',
  EMAIL: 'email',
  ATIVO: 'ativo',
  PAPEL: 'papel',
  PAPEL_ID: 'papel_id',
  EMPRESA_ID: 'empresa_id',
  ID: 'id',
  CRIADO_EM: 'criado_em',
  ATUALIZADO_EM: 'atualizado_em',
} as const

/**
 * UsuariosInclude
 */
export type UsuariosInclude =
  (typeof UsuariosInclude)[keyof typeof UsuariosInclude]
// ...
export const UsuariosExpand = {
  EMPRESAS: 'empresas',
  PAPEIS: 'papeis',
  USUARIOS_FUNCIONALIDADES: 'usuarios_funcionalidades',
  EQUIPE_USUARIOS: 'equipe_usuarios',
  ATENDIMENTOS: 'atendimentos',
  EQUIPES: 'equipes',
  TEMPLATES: 'templates',
} as const

/**
 * UsuariosExpand
 */
export type UsuariosExpand =
  (typeof UsuariosExpand)[keyof typeof UsuariosExpand]
// ...
export type PostUsuarios = {
  /**
   * Nome Usuario
   * Nome único para login no sistema (case insensitive)
   */
  nome_usuario: string
  /**
   * Nome Completo
   * Nome completo do usuário
   */
  nome_completo?: string | null
  /**
   * Nivel
   * Nivel de experiência do usuário
   */
  nivel?: string | null
  /**
   * Telefone
   * Telefone de contato
   */
  telefone?: string | null
  /**
   * Email
   * Endereço de e-mail válido do usuário
   */
  email: string
  /**
   * Senha
   * Hash da senha do usuário (armazenado de forma criptografada)
   */
  senha: Blob | File
  /**
   * Ativo
   * Indica se a conta está ativa e pode acessar o sistema
   */
  ativo?: boolean
  /**
   * Papel
   * Nome papel principal atribuído ao usuário (pode ser nulo)
   */
  papel?: string | null
  /**
   * Papel Id
   * Papeis principal atribuído ao usuário (pode ser nulo)
   */
  papel_id?: string | null
  /**
   * Empresa Id
   * Empresa (tenant) à qual o usuário pertence
   */
  empresa_id?: string | null
}
// ...
export type PatchUsuarios = {
  /**
   * Nome Usuario
   * Nome único para login no sistema (case insensitive)
   */
  nome_usuario?: string | null
  /**
   * Nome Completo
   * Nome completo do usuário
   */
  nome_completo?: string | null
  /**
   * Nivel
   * Nivel de experiência do usuário
   */
  nivel?: string | null
  /**
   * Telefone
   * Telefone de contato
   */
  telefone?: string | null
  /**
   * Email
   * Endereço de e-mail válido do usuário
   */
  email?: string | null
  /**
   * Senha
   * Hash da senha do usuário (armazenado de forma criptografada)
   */
  senha?: (Blob | File) | null
  /**
   * Ativo
   * Indica se a conta está ativa e pode acessar o sistema
   */
  ativo?: boolean | null
  /**
   * Papel
   * Nome papel principal atribuído ao usuário (pode ser nulo)
   */
  papel?: string | null
  /**
   * Papel Id
   * Papeis principal atribuído ao usuário (pode ser nulo)
   */
  papel_id?: string | null
  /**
   * Empresa Id
   * Empresa (tenant) à qual o usuário pertence
   */
  empresa_id?: string | null
}
// ...
export type ApiSuccessGetUsuarios = {
  /**
   * Msg
   */
  msg?: string | null
  meta?: Meta | null
  /**
   * Data
   * Lista de itens do tipo especificado
   */
  data?: Array<AppSchemaUsuariosSchemaGetUsuariosOutput> | null
}
// ...
export type LerUsuariosData = {
  body?: never
  path?: never
  query?: {
    /**
     * All Data
     * Retornar todas as correspondências ou apenas a primeira
     */
    all_data?: boolean
    /**
     * Nome Usuario
     * Nome único para login no sistema (case insensitive)
     */
    nome_usuario?: string | null
    /**
     * Nome Completo
     * Nome completo do usuário
     */
    nome_completo?: string | null
    /**
     * Nivel
     * Nivel de experiência do usuário
     */
    nivel?: string | null
    /**
     * Telefone
     * Telefone de contato
     */
    telefone?: string | null
    /**
     * Email
     * Endereço de e-mail válido do usuário
     */
    email?: string | null
    /**
     * Ativo
     * Indica se a conta está ativa e pode acessar o sistema
     */
    ativo?: string | null
    /**
     * Papel
     * Nome papel principal atribuído ao usuário (pode ser nulo)
     */
    papel?: string | null
    /**
     * Papel Id
     * Papeis principal atribuído ao usuário (pode ser nulo)
     */
    papel_id?: string | null
    /**
     * Empresa Id
     * Empresa (tenant) à qual o usuário pertence
     */
    empresa_id?: string | null
    /**
     * Id
     * Identificador único universal (UUIDv4) da entidade
     */
    id?: string | null
    /**
     * Criado Em
     * Data e hora UTC da criação do registro
     */
    criado_em?: string | null
    /**
     * Atualizado Em
     * Data e hora UTC da última atualização do registro (null se nunca foi atualizado)
     */
    atualizado_em?: string | null
    /**
     * Limit
     * Número máximo de itens a retornar
     */
    limit?: number | null
    /**
     * Offset
     * Offset absoluto para paginação (alternativa a skip)
     */
    offset?: number | null
    /**
     * Cursor
     * Valor do cursor para paginação baseada em cursor
     */
    cursor?: string | null
    /**
     * Campo a ser usado como cursor (padrão: 'id')
     */
    cursor_field?: UsuariosInclude
    /**
     * Direção da paginação por cursor (next|prev)
     */
    cursor_direction?: AppSchemaEmpresasSchemaCursorDirection
    /**
     * Include
     * Atributos a incluir na resposta
     */
    include?: Array<UsuariosInclude>
    /**
     * Exclude
     * Atributos a excluir da resposta
     */
    exclude?: Array<UsuariosInclude>
    /**
     * Expand
     * Relacionamentos a expandir
     */
    expand?: Array<UsuariosExpand> | null
    /**
     * Count Only
     * Retornar apenas a contagem total de itens, sem os dados
     */
    count_only?: boolean
    /**
     * Distinct
     * Aplicar DISTINCT na consulta para evitar duplicatas
     */
    distinct?: boolean
    /**
     * operador da ordem do resultado (ASC | DESC)
     */
    sort_op?: AppSchemaEmpresasSchemaSortOp
    /**
     * ordem do resultado (ex: criado_em)
     */
    sort_field?: UsuariosInclude
  }
  url: '/usuarios/'
}

export type LerUsuariosErrors = {
  /**
   * Erro na requisição
   */
  400: ErrorSchema
}

export type LerUsuariosError = LerUsuariosErrors[keyof LerUsuariosErrors]

export type LerUsuariosResponses = {
  /**
   * Successful Response
   */
  200: ApiSuccessGetUsuarios
}

export type LerUsuariosResponse =
  LerUsuariosResponses[keyof LerUsuariosResponses]